Add the right verion of 1049//咦？今天真是晕了，之前又commit了没保存的版本

+++ 1049-the-sum-of-array-parts.c
@@ -12,7 +12,7 @@ int main()
 	}
 	for(i=0;i<n;i++)
 	{
-		sum+=(n-i)*(i+1)*data[i];
+		sum+=(double)(n-i)*(double)(i+1)*data[i];
 	}
 	printf("%.2lf\n",sum);
 }
